+++ components/src/lib/components/common/CloseButton/CloseButton.stories.tsx
