+++ components/src/lib/index.ts
@@ -1,9 +1,9 @@
-import Title from './components/Title/Title';
-import Button from './components/Button/Button';
-import CloseButton from './components/CloseButton/CloseButton';
-import { ModalBody, ModalFooter, ModalHeader, ModalMain } from './Modal/Modal';
-import Input from './components/Input/Input';
+import Title from './components/common/Title/Title';
+import { ModalBody, ModalFooter, ModalHeader, ModalMain } from './components/Modal/Modal';
+import Input from './components/common/Input/Input';
 import { AlertModal, ConfirmModal, PromptModal } from './ModalPreset/ModalPreset';
+import CloseButton from './components/common/CloseButton/CloseButton';
+import Button from './components/common/Button/Button';
 
 export const Modal = Object.assign(ModalMain, {
   Title: Title,
